deleted home screen, to be replaced by channel_groups_screen
